+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Added
 
 - added date picker for Ausgabe Erscheinungsdatum field
-- added button to inlines that sends the user to the changelist of the items selected in the inline forms 
+- added button to inlines that sends the user to the changelist of the items selected in the inline forms
+- added 'Band' search field to Musiker changelist
 
 ### Changed